Migrate 'count' test
